Back in our webpack.config.js, we need to add these loaders so Webpack knows what to do. Since these aren’t plugins, they go in a separate section: All this does is tell Webpack that if it encounters an imported file ending with .css, it should use the listed loaders to process that CSS file.
Loader order matters for CSS!
Notice how we put css-loader at the end of the array. We must set this order and not the reverse.

Webpack will run the loaders starting at the end, so we want it to read the CSS file into a string with css-loader first, then use style-loader to inject the JavaScript that applies the CSS in that string to the page. It wouldn’t work the same the other way round.